pyproject.toml
[tool.poetry]
name = "papers"
version = "0.1.0"
description = ""
authors = ["Nate Stott <nate.stott@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
